Remove version constraints for plugins required to build tools.
Normally it is not a good idea to place <version> constraints in the
definition of plugins in pom.xml unless a specific version really is
needed.

This patch removes the constraints.

There are various benefits to this, one of them is that it will always
build with the latest packaged version of the plugins on a Debian or
Ubuntu system.

Author: daniel@pocock.pro

R=jia.shao.peng@gmail.com
BUG=

Review URL: https://codereview.appspot.com/122480043

git-svn-id: http://libphonenumber.googlecode.com/svn/trunk@682 ee073f10-1060-11df-b6a4-87a95322a99c