Build the tools and libraries all at once with Maven.
This patch adds a top level pom.xml so it is now possible to checkout
the repository and do

  mvn clean install

and it will build the tools JAR and then the library itself without
requiring any pre-built JARs in the repository

I've also tweaked the Debian packaging to work using maven-debian-helper
and without using Ant or build.xml any more. I'll submit that as a
separate patch.

The patch also removes a reference to oss-parent that appears to be in
java/pom.xml by mistake, this was causing the maven-debian-helper builds
to fail.

Author: daniel@pocock.pro

R=jia.shao.peng@gmail.com
BUG=

Review URL: https://codereview.appspot.com/125530043

git-svn-id: http://libphonenumber.googlecode.com/svn/trunk@687 ee073f10-1060-11df-b6a4-87a95322a99c